revised DataTypes within Character model file. 'age' key is now restricted to a DataType of INTEGER, while 'height' key is restricted to DataType of DECIMAL.

+++ models/Character.js
@@ -19,11 +19,11 @@ Character.init(
         },
      
         age: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.INTEGER,
             allowNull: false
         },
         height: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL,
             allowNull: false
         },
         background_color: {
